requirements.txt
Django>=1.8,<1.9
Django-Select2
MySQL-python
Pillow
Unidecode
amqp
anyjson
billiard
boto
celery
cssselect
dj-database-url
django-celery
django-classy-tags
django-cms>=3.2
django-filer
django-mptt
django-polymorphic
django-reversion
django-sekizai
django-storages
django-treebeard
djangocms-admin-style>=1.0.0
djangocms-column
djangocms-file
djangocms-flash
djangocms-googlemap
djangocms-inherit
djangocms-installer
djangocms-link
djangocms-picture
djangocms-style
djangocms-teaser
djangocms-text-ckeditor
djangocms-video
easy-thumbnails
gunicorn
html5lib
kombu
lxml
psycopg2
pytz
requests
six
tzlocal
xlrd
xlwt
django-countries==3.4.1
